Add a type for formulas not in negation normal form.
This allows formulas to negate a sub-formula, instead of just negating
an atom. The new type is NFormula. Negation normal form formulas will
still be called Formula, so that I don't have to track down everywhere
Formula is currently used, and to encourage the use Formula directly
since it's the good one. It takes less work to get from a (negation
normal) Formula to CNF.