Add a way to customize export.
This adds a new module, BDCS.Export.Customize. The module provides a way
to add new content or override existing content without modifying the
mddb or content-store. The overridden content map is included with the
FSTree and ContentStore during an export.

The content overlay is populated through the use of Customization
objects, which define actions to perform on an FSTree to modify the
exported content. The only Customization type defined for now is
WriteFile, which writes a file.