Switch regex libraries.
Convert from regex-pcre to pcre-heavy. This takes a little more work to
go from a string to a Regex object, but it includes a lot of useful
features like substitutions.

As part of this change, Utils.GetOpt.compilerOpts now uses MonadError
for error reporting instead of ioError, so that the inspect subcommands
can more easily do more than one thing that may produce an error.